remove login and signup page, and made it into one AuthPage
